Grantham distance matrix

A dataset containing Grantham distances in the format of a matrix. These values were obtained from Table 2, Grantham (1974), doi: 10.1126/science.185.4154.862 .

Usage

grantham_distances_matrix

Format

An object of class matrix (inherits from array) with 20 rows and 20 columns.

Source

Table 2, Grantham (1974), doi: 10.1126/science.185.4154.862 .

Examples

grantham_distances_matrix
#>     Ser Arg Leu Pro Thr Ala Val Gly Ile Phe Tyr Cys His Gln Asn Lys Asp Glu Met
#> Ser   0 110 145  74  58  99 124  56 142 155 144 112  89  68  46 121  65  80 135
#> Arg 110   0 102 103  71 112  96 125  97  97  77 180  29  43  86  26  96  54  91
#> Leu 145 102   0  98  92  96  32 138   5  22  36 198  99 113 153 107 172 138  15
#> Pro  74 103  98   0  38  27  68  42  95 114 110 169  77  76  91 103 108  93  87
#> Thr  58  71  92  38   0  58  69  59  89 103  92 149  47  42  65  78  85  65  81
#> Ala  99 112  96  27  58   0  64  60  94 113 112 195  86  91 111 106 126 107  84
#> Val 124  96  32  68  69  64   0 109  29  50  55 192  84  96 133  97 152 121  21
#> Gly  56 125 138  42  59  60 109   0 135 153 147 159  98  87  80 127  94  98 127
#> Ile 142  97   5  95  89  94  29 135   0  21  33 198  94 109 149 102 168 134  10
#> Phe 155  97  22 114 103 113  50 153  21   0  22 205 100 116 158 102 177 140  28
#> Tyr 144  77  36 110  92 112  55 147  33  22   0 194  83  99 143  85 160 122  36
#> Cys 112 180 198 169 149 195 192 159 198 205 194   0 174 154 139 202 154 170 196
#> His  89  29  99  77  47  86  84  98  94 100  83 174   0  24  68  32  81  40  87
#> Gln  68  43 113  76  42  91  96  87 109 116  99 154  24   0  46  53  61  29 101
#> Asn  46  86 153  91  65 111 133  80 149 158 143 139  68  46   0  94  23  42 142
#> Lys 121  26 107 103  78 106  97 127 102 102  85 202  32  53  94   0 101  56  95
#> Asp  65  96 172 108  85 126 152  94 168 177 160 154  81  61  23 101   0  45 160
#> Glu  80  54 138  93  65 107 121  98 134 140 122 170  40  29  42  56  45   0 126
#> Met 135  91  15  87  81  84  21 127  10  28  36 196  87 101 142  95 160 126   0
#> Trp 177 101  61 147 128 148  88 184  61  40  37 215 115 130 174 110 181 152  67
#>     Trp
#> Ser 177
#> Arg 101
#> Leu  61
#> Pro 147
#> Thr 128
#> Ala 148
#> Val  88
#> Gly 184
#> Ile  61
#> Phe  40
#> Tyr  37
#> Cys 215
#> His 115
#> Gln 130
#> Asn 174
#> Lys 110
#> Asp 181
#> Glu 152
#> Met  67
#> Trp   0